requirements.txt
flask==2.3.3
flask-cors==4.0.0
databricks-sql-connector==3.0.2
pandas==2.1.4
gunicorn==21.2.0